Update patch set 4
Patch Set 4:

> looks like this is missing s3transfer
 > 
 > A bigger problem is that they are capping some requirements, mainly
 > jmespath (s3 transfer and botocore are somewhat more topical)
 > 
 > those are my main problems with it.

Thanks for your review. s3stransfer, jmespath and botocore are in boto3's requirement, and will be auto installed when installing boto3 (these libraries will not be imported in the developers' code, only boto3 or botocore will be imported). Do you think it is necessary to add them to the global-requirement?

Patch-set: 4
Label: Verified=0