Update patch set 4
Patch Set 4: Code-Review-1

> > looks like this is missing s3transfer
 > >
 > > A bigger problem is that they are capping some requirements,
 > mainly
 > > jmespath (s3 transfer and botocore are somewhat more topical)
 > >
 > > those are my main problems with it.
 > 
 > Thanks for your review. s3stransfer, jmespath and botocore are in
 > boto3's requirement, and will be auto installed when installing
 > boto3 (these libraries will not be imported in the developers'
 > code, only boto3 or botocore will be imported). Do you think it is
 > necessary to add them to the global-requirement?

No you don't need to add them to global-requirements but they do need to be added to upper-constratints.txt See[1]

What Matthew is referring to is [2] where the requirements for boto3 are capped which can lead to problems *if* other parts of OpenStack use those items.  Both botocore and jmespath *are* already used and within the ranges needed by boto3.  This is probably because they're used by boto which is used by ec2-api (and heat-cfntools, rally and swift3)

I think before we can accept this we'd need to see a discussion about migrating boto users to boto3.  We have old and new libraries in g-r ATM but we try to avoid it if possible.

The -1 is becuase we need to see the discussion


[1] http://git.openstack.org/cgit/openstack/requirements/tree/README.rst#n186
[2] https://github.com/boto/boto3/blob/develop/setup.py#L16-L20

Patch-set: 4
Label: Code-Review=-1