Update patch set 2
Patch Set 2: Code-Review+2 Workflow+1

Normally this would be an outright reject as we don't add items to g-r like this.  However it seem safe in the instance as virtualbmc is only a testing tool used in the gate and we're explicitly adding to to ensure that said testing is co-installable.

Patch-set: 2
Reviewer: Gerrit User 12898 <12898@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2
Label: Workflow=+1