Update patch set 1
Patch Set 1:

> I'm willing to get perspective from the keystone PTL but newton has
 > nee out for > 6 months and this TLS issue has not been *reported*
 > against the newton codebase.  So either consumers are fine with the
 > newton performance or have alternate solutions.
 > 
 > Right now this seems inappropriate for stable/newton

So far, only OpenStack-Ansible has come to us about the issue. I wouldn't be surprised it others have hit this problem in other deployments, but I'd like to know if they have any workarounds that we could document. Ideally, if we could figure that out, I'd be OK documenting those workarounds instead of backporting to Newton for the reasons Tony's listed above (since Newton is in security only maintenance).

Patch-set: 1