Update patch set 1
Patch Set 1:

> We seem to enable using ldap pool by default for ldap backend so we
 > have a feature that is "broken by default".
 > 
 > https://github.com/openstack/keystone/blob/stable/newton/etc/keystone.conf.sample#L1363
 > 
 > And the failure is not that easy to troubleshoot either as the
 > error message is very generic. So the way I see it.
 > 
 > 1. Document the fact that the feature is completely unusable at the
 > current state and advice operators to do either version pinning (on
 > the older version) or disable using ldap pool till upstream have a
 > fix (which we did).
 > 2. Or land this patch and make operators lives easier.
 > 3. Put this feature on the deprecation block if nobody cares about
 > it. But I do remember ldap pool offers measurable performance
 > advantages when it was first implemented. Unfortunately, I am
 > unable to dig up those numbers from the original author right now.

All valid points. I'm OK with option #2 so long as the stable maintenance team is. I just don't want to overstep the boundaries put in place when it comes to release phases. In the event we don't go with #2, then we can document the update and provide some suggestions to operators. I agree that ldappool provides value, so I wouldn't be in favor of option #3.

Patch-set: 1