Update patch set 1
Patch Set 1: Code-Review+2

It looks to me like all the work is already done in that keystone can happily talk TLS to a LDAP server directly it's only when it uses the socket pool (via ldappool) that things go wrong.  The actual fix in ldappool is pretty minimal and clearly fits the bugfix criteria.

ldappool only has a master branch, so we need to be sure that the requirements are compatible with those in stable/newton and they are (yay constraints!)

I'm a little disturbed to see that both UbuntuCloudArchive and RDO are still using ldappool 1.x which may explain the lack of real world consumers hitting this bug.

On balance I think this is appropriate for stable/newton

Patch-set: 1
Reviewer: Gerrit User 12898 <12898@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2