Update patch set 2
Patch Set 2:

> > ya, that's what I though.
 > >
 > > As far as the proper fix goes, the best thing to do would be to
 > > change this commit so that it excludes all the bad releases (look
 > > at things like setuptools as an example).
 > 
 > Sorry I still did not get your point. you mean setuptools for each
 > projects?

The suggestion from Matthew is to do somethign like:

 oslo.middleware>=3.0.0,!=3.1.0,!=3.2.0  etc

So that the first acceptable version is 3.10.0 BUT that's still not acceptable.  The sad fact is that once a stable branch is released we can't correct this types of API incompatibility and we need to rely on upper-constratints.txt to signal the correct version to use.

The reference to setuptools was just an example see [1]

[1] http://git.openstack.org/cgit/openstack/requirements/tree/global-requirements.txt#n411

Patch-set: 2