Remove more linters from global-requirements list
We dropped various linters in https://review.openstack.org/#/c/473094/.
This is a follow-up patch on this.

astroid is only used by pylint and both versions are tightly coupled.
If we force astroid===1.3.8, we cannot install newer pylint like 1.7.1.
It looks good to consider astroid as a linter.

bandit is a linter, so it looks good to drop it.

Change-Id: Ib2e78f6a22b7d90a3304b13228c16d76982fbbef