Update patch set 1
Patch Set 1: Code-Review-1

https://github.com/openstack/ceilometermiddleware/compare/0.5.1...0.5.2

looks like they capped oslo.messaging, this does not match requirements.
https://github.com/openstack/ceilometermiddleware/commit/9f6a8e70a579b5953aa4143e6630291b602c2eff

It looks like the other thing that was broght in was a feature, not sure if it counts as a critical bug or not, but it came in just after newton went into phase two, so may be fine.
https://github.com/openstack/ceilometermiddleware/commit/9f6a8e70a579b5953aa4143e6630291b602c2eff

Patch-set: 1