Update patch set 3
Patch Set 3:

> > upper-constraints needs to be updated first, if the new gabbi was
 > > just released today then it should be in tonights update
 > 
 > Are you saying I need to update upper-constraints in this patch, or
 > wait for the auto job to update them, and then rebase?

You can do either.  In theory gabbi *should* be included in the bot generated change I786d04b1e47fccce6aa9ffdc9538b60f668e9e1d but it isn't

So either something is blocking that (It maybe that something in g-r has a cap in gabbi which is excluding 1.35) or we just got unlucky with the timing.

I'll run the bot locally update this review

Patch-set: 3
Label: Verified=0