Clean up python version conditions
At this point we only care about py27, py34, and py35.  That said, there
are only a few cases where a specific version matters, versus just the
py2/py3 boundary.  With this change set, global-requirements.txt is
updated in such a way that future changes to supported minor versions of
python won't cause it to be out of sync.  (That is, we shouldn't have to
do this again when we change to py28/py36.)

Change-Id: I50c21c5c8537c810b42682bdf2198cf6c22c5941