Update patch set 6
Patch Set 6: Code-Review-1

After discussing with gcb, we're going to blacklist the 2.19.1 version and he's going to revert https://review.openstack.org/#/c/327564/ and redo it but with a UserWarning so we at least have a cycle of warnings before making that change which could break things. We don't really want to release that on the world right before feature freeze.

Patch-set: 6
Label: Code-Review=-1