Update patch set 1
Patch Set 1:

Tempest is branchless and only has master. So the requirements for tempest follow the global requirements repository for master. Because of this tempest is installed in a venv to isolate it from the rest of the world.

This actually reminds me that the issue here is likely being caused by the projects using the TEMPEST_ALL_PLUGINS devstack-gate flag. This uses a venv with system site-packages enabled and there are a lot of weird cross-interactions between requirements that occur when you do this. Especially across release boundaries. I looked at this patch before morning coffee so I didn't even consider that. (although this is still a good patch, it just might not fix the underlying cause of the mismatch)

The real fix to prevent this kind of confusion is to get all these projects' tempest plugins off of TEMPEST_ALL_PLUGINS and into a separate branchless repo that can be properly venv isolated. Which will happen in the next cycle as part of https://governance.openstack.org/tc/goals/queens/split-tempest-plugins.html (the specific steps are https://etherpad.openstack.org/p/tempest-separate-plugin )

Patch-set: 1