Update patch set 1
Patch Set 1: Code-Review-2

Procedural -2.

We're now in requirements freeze for the Pike release of OpenStack.  If you feel that this chnage is critical please apply for a Feature Freeze exception.

To do this please email openstack-dev with the rationale with '[requirements][ffe] in the subject.

I get the desire to include this and I'm generally supportive but we need to consider things like do we also bump g-r or is 4.4.0 okay?  If we do bump g-r that'd mean re-releasing openstackclient.

So we need to have that discussion on the mailing list

Patch-set: 1
Label: Code-Review=-2