Update patch set 3
Patch Set 3:

Procedural -2:

We're now in requirements freeze for the Pike release of OpenStack.  If you feel that this chnage is critical please apply for a Feature Freeze exception. by email openstack-dev with '[requirements][ffe]' in the subject.

Sorry my script didn't apply this comment when it dropped the -2

Patch-set: 3