Update patch set 1
Patch Set 1: Code-Review-2

Procedural -2:

We're now in requirements freeze for the Pike release of OpenStack.

This can wait until we open for queens.  I know that's not exactly what I said on IRC but I've had time to consider it and I can't justify this for unit tests on OSX.

Patch-set: 1