Update patch set 4
Patch Set 4: Code-Review-2

Procedural -2:

Adding a new requirements is pretty low risk and this one especially so as it's just for documentation.  Is it actually going to be used on stable/pike ?  Once we switch to zuulv3 we'll clearly have the per-repo jobs but we'd need to backport the (probably trivial) docs changes for this to be used.  Is there value in generating these docs on stable/pike?

Patch-set: 4