Update patch set 4
Patch Set 4: Code-Review+2

tonyb - in v3 jobs can be defined in all branches, so I think it would be nice to have the ability to backport it if we want to. We're also using shade as one of our early repos for test development, and shade follows g-r, so it would be nice to get it in to g-r so we can get those documented.

Patch-set: 4
Reviewer: Gerrit User 2 <2@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2