Update patch set 2
Patch Set 2: Code-Review-2

This isn't appropriate for pike.

At the very *least* the Ic70024707d1ea1b22e678dd9268485a9485204f4 will trigger a new-release of monascaclient (1.8.0) which would then need to go through the release *and* requirements FFE process which means that this review is broken by design.

Sorry it's just too late in the cycle for this change.

You're free to disagree and apply for the appropriate FFEs on the os-dev mailing list but you have until we start tagging RC1s to get that done.

Patch-set: 2
Label: Code-Review=-2