Update patch set 1
Patch Set 1: Code-Review-2

Procedural -2:

We're now in requirements freeze for the Pike release of OpenStack.  If you feel that this chnage is critical please apply for a Feature Freeze exception. by email openstack-dev with '[requirements][ffe]' in the subject.

Having said that this is only used by heat-agents which doesn't use constraints so I'm not sure it's worth applying for an FFE

Patch-set: 1