Update patch set 1
Patch Set 1: Code-Review-2

Procedural -2:

We're now in requirements freeze for the Pike release of OpenStack.  If you feel that this chnage is critical please apply for a Feature Freeze exception. by email openstack-dev with '[requirements][ffe]' in the subject.

(switching from Matthew's -2 to mine in case these need to get lifted)

Patch-set: 1