Update patch set 1
Patch Set 1:

Please answer the following questions in the commit message.

https://github.com/openstack/requirements/#for-new-requirements

We are in a requirements freeze, so if this needs to get into pike you'll need to mail the mailing list for a FFE (feature freeze exception) with the subject line [FFE][requirements].  I'll let tonyb do the -2 though.

I'd also like to see a general discussion on the mailing list about migrating away from testrepository to stestr openstack wide in queens.

Patch-set: 1
Label: Verified=0