Update patch set 1
Patch Set 1:

> Stable Ocata is only open for critical security bugs.

I need to admit I was wrong here.  I was thinking about newton.  Ocata is open for bugfixes.  That's a small but important distinction.

We need to show that 2.5.0 is somehow broken *and* 3.5.0 is the next version that fixes the issue.  That would mean excluding 3.[01234].* on master and then backporting that to ocata *or* explaining why that isn't needed.  Assuming that happens you'd also need to place a global ban on those versions in global-requirements.txt

Given that master is frozen now I doubt that will happen before the release.  Which means you'll need to backport from master -> stable/pike -> stable/ocata

I suspect that the *package builds* for Red Hat OpenStack will just need to use 3.2.0.

I'm still not sure how the existence of these packages requires this change.

Patch-set: 1