Update patch set 2
Patch Set 2:

Procedural -2:

We're now in requirements freeze for the Pike release of OpenStack.  If you feel that this change is critical please apply for a Feature Freeze exception. by email openstack-dev with '[requirements][ffe]' in the subject.

Having said that I think the removal will have to wait for queens as it's used beyond the packages you mentioned.

 Package      : positional [positional>=1.1.1] (used by 5 projects)
 Re-Release   : 5 projects
 openstack/keystoneauth                        []
 openstack/keystonemiddleware                  []
 openstack/oslo.context                        []
 openstack/python-keystoneclient               []
 openstack/python-octaviaclient                []


There are also other users outside of requirements management but they don't impact this.

Patch-set: 2
Label: Verified=0