Update patch set 1
Patch Set 1: Code-Review-2

Procedural -2:

We're now in requirements freeze for the Pike release of OpenStack.  If you feel that this change is critical please apply for a Feature Freeze exception. by email openstack-dev with '[requirements][ffe]' in the subject.

As you're trying to exclude the version in constraints you also need to update that to set the constrained version to ===0.4.17

Doing that change will impact the following projects. 

 Package      : pyroute2 [pyroute2>=0.4.17] (used by 8 projects)
 Included in  : 3 projects
 openstack/neutron                             [cycle-with-milestones]
 openstack/octavia                             [cycle-with-milestones]
 openstack/tacker                              [cycle-with-intermediary]
 Also affects : 5 projects
 openstack/kuryr                               [None]
 openstack/kuryr-kubernetes                    [cycle-with-intermediary]
 openstack/networking-bagpipe                  [cycle-with-milestones]
 openstack/neutron-fwaas                       [cycle-with-milestones]
 openstack/vmware-nsx                          [None]

Neutron, Ocatavia and Tacker aren't big problems but the other 5 are harder.

Over all I think it'd be best to wait for queens and then set the pyroute2>=0.4.19

Patch-set: 1