[WiP] Another one of tonyb's silly ideas.
The purpose of this tool is to append a templated string to a commit message
in the scenario that the requirements repo has branched but the project
repo has not.  This could indicated that the proposed changes are destined
for the next release of OpenStack and shoudl be reviewed carefully.

We'd need to talk to infra about this as we can *in theory* use this
tool in propose-updates.  It doesn't actually fix anything so I'm not
convinced that it's worth it.

Change-Id: I9874e801f2e7ac511d1f3a7a75467c7747b14ddd