Update patch set 1
Patch Set 1:

Anyone that wants to validate the python packages environment, before or after installing the package, would be interested about this.

If we don't provide a pip version that has `check` on it they will not be able to use it.

A common use case is to include "pip check" as a command on tox, just before running the testing commands. This validates that there are no missing dependencies or conflicts.

Patch-set: 1
Reviewer: Gerrit User 24162 <24162@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0