Add PyJWT to global requirements
JWT is a standardized bearer token format[0][1]. Keystone will be
adding support for JWT tokens as an alternative to Fernet
(which is effectively unmaintained). PyJWT is the implementation
of JWT for python.

The consumer for PyJWT will be Keystone.

PyJWT is MIT licensed [2].

PyJWT looks to be well written.

PyJWT is actively maintained (last update as of this commit message was
8 days ago).

PyJWT is Python3 compatible (3.3, 3.4, 3.5, 3.6)[2]

PyJWT is packaged for Fedora[3] and Ubuntu[4].

PyJWT functionality is not easily replicated by another library in
global-requirements. It is the expected Python implementation of
JWT functionality.

Currently PyJWT only requires Cryptography >= 1.4 as a runtime
dependency.

[0] https://tools.ietf.org/html/rfc7519
[1] https://tools.ietf.org/html/rfc7797
[2] https://pypi.python.org/pypi/PyJWT/1.5.3
[3] https://www.rpmfind.net/linux/rpm2html/search.php?query=python-jwt
[4] https://packages.ubuntu.com/source/xenial/pyjwt

Change-Id: Ie192823572bc5b0d1d290853920afeae80217a53