Raise castellan minimum to 0.13.0
Commit 8980bf7da55dd084ad84c84534fe937f0d43b9c0 in Castellan deprecated
config option key_manager.api_class, switching to use key_manager.backend.

In order for consumers to switch to using backend, we need to ensure the
version of castellan used is at least 0.13.0 to have this option present.

This is causing at least Cinder docs builds to fail due to treating
warnings as errors and getting deprecation warnings during the build.

Related-bug: #1718468
Change-Id: I112f7a4654a65c3291526b408838d7b1c175b069