Update patch set 1
Patch Set 1:

tony, are you saying that the requirements bot update would need to be able to change the values in consuming projects so that we are consistant on this openstack wide?  If so, I agree.

Patch-set: 1
Reviewer: Gerrit User 14288 <14288@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=0