Update patch set 1
Patch Set 1:

> tony, are you saying that the requirements bot update would need to
 > be able to change the values in consuming projects so that we are
 > consistant on this openstack wide?  If so, I agree.

No I'm saying that upper-constratints.txt is a generated file.  Changing the data will just get clobbered the next time the bot runs.

So when we make this change we also need to update the tool that generates the file so that we don't loose these chnages 24 hours after merging them.

The optional step is ensure that all projects use correctky canonicalized names in requirements files but that isn't strictly needed and if we ever land the per-project requirements that will go away.

Patch-set: 1