Update patch set 1
Patch Set 1:

I think that the requirement for having a product already using pip 9 before adding it to reference requirements is unreasonable because it creates a paradox.

No product will be able to start using "pip check" validation (pip>=9) because the reference pip does not support it. 

Please reconsider this. Note: obviously that I could easily create a POC CR on any of the products that is using the new pip features, but it will fail to be merged because it would use a newer pip than the one specified in requirements.

Patch-set: 1