Update patch set 4
Patch Set 4:

Can this job be used as is for other projects?

if not shouldn't we do this work in openstack-zuul-jobs so it can be shared in common, and replace the legacy-requirements job?

Also should we make project-requirements-change.py a proper script from this module rather than sticking it in plybooks/files ?

Patch-set: 4
Label: Verified=0