Add RSD client into requirements
RSD client has released the first version of its OpenStackClient
plugin. Since we would like to add this plugin into OpenStackClient
doc, we need rsdclient to be part of this list so that other projects
could add it into the requirements.txt.

Questions for new requirement:

1. Is the library actively maintained?

Yes, python-rsdclient released the initial version in this month and
will continue to support further RSD version in the future.
https://github.com/openstack/python-rsdclient

2. Is the library good code?

This library followed the OpenStack convention to add all necessary
unit test cases, and the OpenStackClient plugin provided by this
project has passed related OpenStackClient tests.

3. Is the library python 3 compatible?

Yes, it is compatible with python 3, and passed all python 3 test
cases.

4. Is the library license compatible?

Yes, it released under Apache 2.0 license.

5. Is the library already packaged in the distros we target (Ubuntu
latest / Fedora latest)?

Current it has not been packaged in any distros, because it just
released the initial version 0.1.0.

6.Is the function of this library already covered by other libraries
in global-requirements.txt?

No, this library provides the specific RSD OSC plugin, which it
reason why this library is created.

7. Is the library required for OpenStack project or related dev or
infrastructure setup? (Answer to this should be Yes, of course) Which?

Yes, when add the RSD plugin into OpenStackClient plugin list, it
requires it has been add global requirements at first, then OSC will
install it to automatically generate usage document.
https://review.openstack.org/#/c/511683/

8. If the library release is managed by the Openstack release process
does it use the cycle-with-intermediary release type?

This project is not released by the OpenStack release project, because
it is not official OpenStack project, but only provides specific RSD
plugin to allow user to invoke RSD API through OpenStackClient.

9. Do I need to update anything else?

No, the initial version also has been added into upper-constraints.txt.

Change-Id: I156cdfbd0afaf9f7e5f41749545e2b8bff663b96