Add copy project-config's legacy-requirements-cross-* jobs
These are going to be rewritten to native zuulv3 jobs
in the next commit.

Needed-By: I3b0dbdcc02a0a8cff0c52701fadf7632648afb97
Change-Id: I3b9a86528350242c4d785d0f49c318de6c29ec75
Closes-Bug: #1730673

Add requirements-check job

This is a straight copy of the original gate-{name}-requirements job,
except it copies the script over to the host rather than depending on a
baked-in copy of the script. Followups should rework it to stop using
zuul-cloner.

(jlk) updated to fix pep8 errors so that the patch can land.

Change-Id: I734db3b3e4d445e15084d1ca6e60cf2fe392385d

Fix check-requirements

Fix check-requirements: The assumptions with remote branch are not valid
anymore with Zuul v3, the change is on the proposed branch, check against
HEAD^1 instead.

Use the command module instead of script to get stdout in the
job-output.txt.gz file. Zuul v3 currently will not record the output from script
module in that file, only when command is used.

Note that the script module copied the file to the remote system - the
system where the tests are run - while command does not do this. But
the command is already on the remote system, use correct path for it.

Co-Authored-By: James E. Blair <corvus@inaugust.com>
Change-Id: Ib44332b4daf63b9d3fca6eadf0b4825614ca4aeb

Add a Zuul v3 native cross-check job

Convert the existing legacy-cross-* jobs over to the
new cross-check-job.

Adds a small bump to upper-constraints in order to test
the success of this review.

Co-Authored-By: Dirk Mueller <dirk@dmllr.de>
Change-Id: Iba6a025e3a10185ad1d8217244b967713cace3f6