Update patch set 6
Patch Set 6:

I only half understood why this was needed and don't have the understanding to verify it isn't needed anymore.

So if and infra-core can confirm or if we can get a noop constraints change to validate that cross-nova-functional still work on top of this that'd be great.

Patch-set: 6
Label: Verified=0