Add pycodestyle
pep8 was renamed to pycodestyle, so allow that one in order to
be able to transition to it.

The depends-on remove pyflake mentions as most projects adopted
flake8/hacking, we don't need to allow explicit dependencies
on the tools integrated by flake8 (mccabe, pycodestyle/pep8, pyflakes).

Cleanup those from the tests/ files to avoid grep'ing finding them.

Depends-On: I12b5e114c5c6783e9da5cca2818ac9266a00fc87
Depends-On: I4c941c9f35865b70360b7ed0c42af8b2b042edfb
Depends-On: Ib145e2afa97f441c07fada9c30f0f0e2410870ae
Depends-On: I99ad1c1c31c4ba9e0c7e3ed692e650a9b4aba13b
Depends-On: Id44c90ddd4a8a2204fd7ce600097bafbc9469993

Change-Id: If9f30bf39143d5e6fefca794385cafc4456d2197