Update patch set 2
Patch Set 2:

Matthew, I posted reviews to remove pep8 for neutron-lib (and others) and added pep8 back on this one. so far this is about *adding* pycodestyle (which should not be controversial) and about removing pyflakes. Do I get correctly that your -1 is about removal of pep8 (which is no longer in revision 2)? or is this about removal of pyflakes?

Patch-set: 2
Label: Verified=0