Update patch set 10
Patch Set 10:

Matthew, what exactly do you mean with dependent commits? Best to discuss on IRC.

In general: With Zuul v3, you can depends-on a change - even for a new job - and test that it works. And that's what I did here to test that the change does the right thing. I tested both that it failed when it should fail and that it succeeded when it should...

Patch-set: 10