Update patch set 2
Patch Set 2: Workflow-1 Code-Review-2

I don't think we should update test-requirements.txt in the same commit.

Also, this is ocata, and as this is not fixing a HARD break or a sec issue we should not change GR like this.  Also, we don't update the >= or <= or > or < values for global-requirements after the branch becomes stable without GOOD reason.  We do allow != values to be added though (and UC updates) as long as they follow the stable policy.

Patch-set: 2
Reviewer: Gerrit User 14288 <14288@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-2
Label: Workflow=-1