Update patch set 2
Patch Set 2:

um... i'll be honest, i'm only fixing this because i'd like to remove numpy from upper-contraints which *is* breaking ceilometer's gate currently.

i basically just backported the patches that fixed original issue in pike but if you have a better solution, go for it. as i see it, it's either this or we cap docutils and i believe upping the lower bound is better than a cap. either way, ocata doc check needs to be fixed. imo, this should've been backported long ago but it seems like the gate was never running docs and was broken this since early this year.

Patch-set: 2