Update patch set 1
Patch Set 1: Code-Review+2

I am positive about defusedxml, as it is a more secure option to adopt for openstack. that said I'm a little bit worried about the number of XML implementations that we have already in place. I'll need to go through the list. still +2 for now.

Patch-set: 1
Label: Code-Review=+2