Update patch set 6
Patch Set 6:

okay, so this actually fixes wheel builds for both ubuntu and centos. I'd like proceed about getting this merged.

Once merged, the next step is to rework the job so that we can actually build wheels when somebody proposes a change to bindep.txt and allow us to iterate faster on changes.

This also makes it easier to bring only opensuse and fedora for example.

Patch-set: 6