Update patch set 3
Patch Set 3:

FYI, here is the feedback of the protobuf developers in their bug regarding this:

https://github.com/google/protobuf/issues/3942

Sorry for the breakage. The 3.5.0 release had regression on the service reflection which was fixed in 3.5.0.post1. Please depend on that instead.

I'd recommend specify dependency as version range on the minor version. in this case: "protobuf>=3.5.0,<3.6.0".

Patch-set: 3
Reviewer: Gerrit User 13252 <13252@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0