Add cliffdemo
cliff uses cliffdemo as a requirement for its tests and its doc builds.
Up until now it has been included as a sidecar via a line in tox.ini ...
but with the move away from using tox for documentation builds, we're in
a pickle with how to install cliffdemo.

This adds it to global-requirements so that we can use normal
requirements declarations to indicate it's necessary.

Change-Id: I593105771611412c51c70e13a99df9e2acd3c186