Update patch set 2
Patch Set 2: Code-Review+2

I think this is probably good for now until we have a better understanding of the root cause. It might be that we just want it to be <1.20, but until we know the cause and when it was introduced, I think the safest route is to just go back to our previous constraint.

Patch-set: 2
Label: Code-Review=+2