Update patch set 4
Patch Set 4:

To my understanding not everything that hits their Travis CI is released as such. I believe only items which are tagged as a release are released to PyPi, there seems to be discipline when tagging releases that all tests pass when reviewing the Travis history.

Patch-set: 4
Label: Verified=0