Add raven
Raven is client for Sentry error tracking tool. It is required
for sentry integration into oslo.log module.

Is the library actively maintained?
Raven is actively maintained (https://github.com/getsentry/raven-python)

Is the library good code?
Code looks good to me, testing via tox included.

Is the library python 3 compatible?
Yes

Is the library license compatible?
Yes (BSD licence)

Is the library already packaged in the distros we target 
(Ubuntu latest / Fedora latest)?
Yes, both.

Is the function of this library already covered by other libraries
in global-requirements.txt?
No, as far as it's the only python library to integrate Sentry.

Is the library required for OpenStack project or related dev or
infrastructure setup? (Answer to this should be Yes, of course) Which?
Sentry integration improves develepers experience and allows devs to
collect exceptions and bugs from live environment.
https://review.openstack.org/#/c/518949/
https://blueprints.launchpad.net/oslo.log/+spec/sentry-handler

If the library release is managed by the Openstack release process
does it use the cycle-with-intermediary release type?
Raven is 3rd party library, has nothing to do with OpenStack release
process.

Do I need to update anything else?
Nope.

Change-Id: I292b0331a67f238779c8d565c5288d59a34dac69