Update patch set 6
Patch Set 6:

Matthew, I agree the cap is ugly. My current suggestion would be to wait for some response from the kubernetes site. Innthe worst case we have to remove the cap again, but for now everything seems to pass with the older version and we're not explicitly requiring the newer one

Patch-set: 6