Update patch set 1
Patch Set 1:

I would agree Fedora should probably not package that version, or create the directory.

But see linked issue for the fix that should be included in the next release of rtslib. The way they are handling this in 2.1.64 is not safe so we should blacklist it.

Patch-set: 1
Label: Verified=0