Add uwsgi to upper-contraints
As most of the projects switch to wsgi based deploys, the usage of uwsgi
is going up alot more. Glance recommends using 'uwsgi' to deploy[0].
That fact alone makes this an indirect dependancy in my opinion.

The LOCI images also pull in uwsgi by default, and that is the only
package we pull in that is not in upper-constraints.txt.

[0] https://docs.openstack.org/glance/pike/admin/apache-httpd.html#uwsgi

Change-Id: I56fd8f855c87aebade49f4f491c0ba63991efba6