Update patch set 3
Patch Set 3: Code-Review-1

the failing tempest test are passing for me locally so i think this might be an issue in the gate due to limmit on how many openfiles we can have.

looking at log stash for the last 90days
http://logstash.openstack.org/#/dashboard/file/logstash.json?query=message:%5C%22*Too%20many%20open%20files*%5C%22%20AND%20tags:%5C%22screen-n-cpu.txt%5C%22
the first time this error was reported was in 
https://review.openstack.org/#/c/484386/1
so either using pyroute2 increased the amount of open file descriptors we are using or we are leaking file descriptors.

if we are leaking the descriptors it could be either in os-vif iteslf or in the pyroute2 lib we are now pulling in.

Patch-set: 3
Reviewer: Gerrit User 11604 <11604@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1