Update patch set 3
Patch Set 3:

A link to a bug that would explain why we want to block 1.8.0 would be handy. Without it, I can't assess if we need to block it, or just bump u-c. Also, if you want to consume new stuff from the library in nova or neutron, you may need to bump minimal version in g-r. (not sure if that's the case.)

Patch-set: 3