Update patch set 1
Patch Set 1:

Just fyi, that change is already in the master, so likely things will continue broke in future sphinx releases.

Here is how we fix it in octavia, the key is to make meaningful __init__.py not empty:
https://review.openstack.org/#/c/534765/


Add license for empty __init__.py

This also fix build-openstack-sphinx-docs, there was a change introduced
in sphinx 1.6.6:
https://github.com/sphinx-doc/sphinx/pull/4335/files

If the size of __init__.py is less than 2, then the module would be
skipped which will cause the sphinx consistency checking failing later.

Patch-set: 1