Added transitions library to masakari intrusive-instance-monitoring
As the first step to get these defined in requirements.txt,
adding to global-requirements and upper-constraints.

* Is the library actively maintained?

Yes, there are 38 contributors and all source is available on github.

* Is the library good code?

Code is well documented and tested.

* Is the library python 3 compatible?

Yes.

* Is the library license compatible?

Yes, it is using the MIT license.

* Is the library already packaged in the distros we target?

Yes, I got it from Ubuntu latest

* Is the function of this library already covered by other libraries
in global-requirements.txt?

* Is the library required for OpenStack project or related dev
or infrastructure setup?

Masakari

* Do I need to update anything else?
No

Change-Id: I2f6e4abd248bf2b05fca5af7364e32d4fe452c31
Implements: blueprint intrusive-instance-monitoring