Block construct >= 2.9 and pin to 2.8.21
the release 2.8.22 of 'construct' lib has introduced a breaking change
to Python API wrt to available exceptions (see Issue #453 in construct),
which broke ironic-instpector and currently blocks its CI.

Current plan according to construct maintainer is to pull 2.8.22
from PyPI and move this change to 2.9 series.

This patch reverts the u-c for this lib to 2.8.21, and modifies the
g-r accordingly.

Change-Id: I5c63c7f0b5d3122c0738d8591352f580185bd0d6