Block construct >2.8.21
the release 2.8.22 of 'construct' lib has introduced a breaking change
to Python API wrt to available exceptions (see Issue #453 in construct),
which broke ironic-instpector and currently blocks its CI.

The only openstack project seeming to depend on construct
is ironic-inspector, and it has plans to drop this dependency.

To fix the CI in the meantime, this patch reverts the u-c for construct
to 2.8.21, and modifies the g-r accordingly.

Change-Id: I5c63c7f0b5d3122c0738d8591352f580185bd0d6