Update pylxd to 2.2.4 upper constraint as fix for lxd 2.0.11
This fix to the stable/ocata upper-constraints.txt is to ease pylxd
to 2.2.4 from 2.2.2.  This is to fix Bug #1746068.  It causes the Zuul
gate job "legacy-tempest-dsvm-lxd-ovs" to fail which means nothing can
be backported to nova-lxd for stable/ocata.

lxd 2.0.11 grew a new feature ('description' tag for a container) and
this has been added to the stable Ubuntu Xenial 16.04 release.
Unfortunately, pylxd 2.2.2 can't handle this extra attribute and crashes
out.  This was fixed in pylxd 2.2.4.

An example of a failure is: https://review.openstack.org/#/c/538210/

Change-Id: Icf315725e6e5099b656ad56c3d98f9a466835439
Related-Bug: #1746068