Update patch set 1
Patch Set 1:

Matthew, thanks for the consideration.
I am not strongly opinionated towards the lower bound or the upper bound.
Given the incompatibility we found in functional tests, I'm more inclined to reset upper constraints to 0.9.19. If there is a compelling reason to raise the lower bound, I'm okay with it too. Just want to make sure we are not allowing this version range where things behave differently.

I'd prefer leaving the decision to Monty. He has spent a HUGE effort on SDK improvement. He knows better which version we should use. It was a big surprise to me that after all those changes, we only found two small issues. :)

Patch-set: 1