Update patch set 2
Patch Set 2: Code-Review-1

The plan is to migrate the openstack actions out of mistral into mistral-extra. I know that right now it is not being used, but we were a bit late and had to wait for the queens release. There's some cards to add functionality to it in this cycle.

Patch-set: 2
Reviewer: Gerrit User 15895 <15895@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1