Remove cap on websocket-client
The new version contains support for Kubernetes core Workloads API.

To update to this version we need to:
a. Remove the cap on websocket-client
b. Alter the constrained version of websocket-client - to be co-installable
with openstack libs (python-zunclient) and kubernetes
c. Raise the minimum acceptable version fro kubernetes - to be compatible
with the wider range of websocket-client

a follow-on commit to do b and c after zuneclient is release will follow

Change-Id: I8f65ca4ffc1cab7a7eb0708884ee311a6aeb0e26
Signed-off-by: Matthew Thode <mthode@mthode.org>