Drop tempest-lib from g-r
It has been deprecated for some time already and has
seen its last release in 2016, so I'd say it is time
to cleanup the cruft.

Deprecation notice is the last patch that merged
into that git repo - see
https://github.com/openstack/tempest-lib/blob/master/README.rst
Everything else get's -2'ed and abandoned. so I don't think
it is a good dependency for OpenStack anymore.

Closes-Bug: #1753969
Depends-On: Ie63ac20e452cd6cde6cbb0e629d8d8c8fb5adf38
Change-Id: Ib4d550870bbdb85fb2a9df20d295015dd3f03597