Update patch set 2
Patch Set 2: Verified-1

This change depends on a change to a config project.

The syntax of the configuration in this change has been verified to be
correct once the config project change upon which it depends is
merged, but it can not be used until that occurs.

Patch-set: 2
Reviewer: Gerrit User 22348 <22348@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=-1