Update patch set 2
Patch Set 2: Code-Review+1

Looks good to me, my only concern is that now we're effectively treating horizon and neutron as cycle-with-intermediary deliverables and I'm not sure that's been communicated.

Of course we can gloss over that from a CI POV but it does make life strange for pypi consumers.  As long as we have a plan .... do we have a plan? :D

Patch-set: 2
Label: Code-Review=+1