Update patch set 1
Patch Set 1:

I think the question is if requirements (specifically global-reqs.txt) should be descriptive or prescriptive.

If prescriptive, we can set this as this is what we want to limit support to in openstack.
If descriptive, we do not set this as upstream officially supports py3.

I'm more inclined to be descriptive.  (so won't change my vote, but I won't block it).

Patch-set: 1