Update patch set 1
Patch Set 1: Code-Review-1

So we're in a tight situation here.

The recent ovsdbapp release was from queens which obviously applies to master as well ... I'll do some cherry-picking and rebaseing.

That hard part is it's high priority bug but we don't want to cause more brakage so we have the balance speed with risk :(

Patch-set: 1
Label: Code-Review=-1