Add some OpenStack networking-* projects to global-requirements (part 1)
Now that we're publishing them to PyPI we can begin the work of
eworking the plugin projects to depend on the PyPI versions instead
of listing direct git url depends. To do that, we need them in
global-requirements.

Add networking-odl, networking-bagpipe, networking-bgpvpn,
networking-sfc,networking-l2gw,neutron-lbaas,neutron-fwaas, neutron-vpnaas,
and neutron-dynamic-routing.

Add these only to global-requirements for now - that way
tools/tox_install.sh continues to work. As next step we can migrate from
tox_isntall.sh to requirements and then add the constraints.

Change-Id: I9feeda89b87bd88636170fee25dc6d1c20208e37