Update patch set 1
Patch Set 1:

> Let me fix that for you properly...

The changes you propose will prevent a project from depending on the master branch of a project. It happens that many networking-* or neutron-* project are co-developed with neutron and sometimes will depend  on features that are not in a released version of neutron. Your changes would, unless I misunderstood, prevent this workflow.

Perhaps the current practice has drawbacks and needs to be revisited, although it seems to me quite legitimate to expect that to be possible in the Neutron stadium.

Either way, I think that it would be preferable to first discuss and agree on a solution that will work for everyone, rather than rush on this path.

Patch-set: 1
Reviewer: Gerrit User 12021 <12021@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0