Bump ryu to 4.18 to avoid pip import issues
Importing pip in code is apparently not supported, but the ryu
library is using it internally. Version 4.18 included a change
to only import pip where needed in a specific method that we
appear to not call from any OpenStack code. Bumping u-c to pick
up this change and test with affected projects.

Pip issue with more details:
https://github.com/pypa/pip/issues/5081

Depends-on: https://review.openstack.org/554053/

Change-Id: Id0a1ff17ef7e2f1dcfb607cfc62bd12862205172