Update patch set 9
Patch Set 9:

> Can you answer the questions here?
 > 
 > https://github.com/openstack/requirements/#for-new-requirements

Is the library actively maintained? Yes.
Is the library good code? Yes.
Is the library python 2 compatible? Yes.
Is the library python 3 compatible? Yes.
Is the library license compatible? Yes (Apache 2).
Is the library already packaged in the distros we target (Ubuntu latest / Fedora latest)? Not at the moment.
Is the function of this library already covered by other libraries in global-requirements.txt? No.
Is the library required for OpenStack project or related dev or infrastructure setup? Yes.
Which one? Dragonflow (https://review.openstack.org/#/c/518670/).
If the library release is managed by the Openstack release process does it use the cycle-with-intermediary release type? Independent.
Do I need to update anything else? No.

Patch-set: 9
Label: Verified=0